+++ packages/rs-drive/src/drive/identity/balance/mod.rs
@@ -1,2 +1,4 @@
 mod fetch;
 mod update;
+mod prove;
+mod queries;
