+++ packages/rs-drive/src/drive/identity/fetch/prove/mod.rs
@@ -1,3 +1,4 @@
 mod full_identities_by_public_key_hashes;
 mod full_identity;
 mod identity_ids_by_public_key_hashes;
+mod balance;
